blackoilnewtonmethod.hh: rename to blackoilnewtonmethod.hpp
